refactor DB optimization code to namespace and subscriber
